requirements.txt
#Database connnect
gsheetsdb==0.1.13.1
google-auth==2.16.0
shillelagh==1.1.5

#Authenticator
PyYAML==6.0
bcrypt==4.0.1

# Streamlit
streamlit==1.19.0
streamlit-authenticator==0.2.1
streamlit-option-menu==0.3.2
plotly-express==0.4.1
geopandas==0.12.2
streamlit-lottie==0.0.3
openpyxl==3.1.2
altair==4.2.2
matplotlib==3.7.1
numpy==1.24.3
pyxlsb==1.0.10
XlsxWriter==3.1.0
squarify==0.4.3
